backend/requirements.txt
# Flask et extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Base de données
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Sécurité
bcrypt==4.1.1
python-dotenv==1.0.0
PyJWT==2.8.0

# HTTP client pour Mayan et Ollama
requests==2.31.0
httpx==0.25.2

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Utilitaires
python-dateutil==2.8.2
gunicorn==21.2.0

# Redis pour sessions et rate limiting
redis==5.0.1

